Styling Fix and Code Refactoring
