+++ src/screens/home/Home.tsx
@@ -1,7 +1,7 @@
 import Profile from "./components/Profile.tsx";
 import Skills from "./components/Skills.tsx";
 import Experience from "./components/Experience.tsx";
-import ProjectList from "./components/ProjectList.tsx";
+import Projects from "./components/Projects.tsx";
 
 const Home = () => {
     return (
@@ -9,7 +9,7 @@ const Home = () => {
             <Profile />
             <Skills />
             <Experience />
-            <ProjectList />
+            <Projects />
         </div>
     );
 };
